+++ python/cpp/pytext.cpp
@@ -20,7 +20,6 @@
 #include <stdio.h>
 #include <nvstrings/NVStrings.h>
 #include <nvstrings/NVText.h>
-#include <nvstrings/util.h>
 
 // utility to deference NVStrings instance from nvstrings instance
 // caller should never dextroy the return object