+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Version jump from 0.3->0.7 is to align with other RAPIDS projects.
 
 ## Improvements
 
+- PR #236 Removed internal util.h from include dir
 - PR #216 Fixed build instructions and removed obsolete doc/build files
 - PR #178 Added compute_statistics to C++ interface
 - PR #173 Added flake8 checks to for /python files